requirements.txt
python-dotenv
openai
streamlit
gspread 
google-auth
google-api-python-client